Add a support/xml artifact for common substrate support #294
