+++ extensions/aws-eks/runtime/pom.xml
@@ -51,7 +51,7 @@
         </dependency>
         <dependency>
             <groupId>org.apache.camel.quarkus</groupId>
-            <artifactId>camel-quarkus-core-xml</artifactId>
+            <artifactId>camel-quarkus-support-xml</artifactId>
         </dependency>
         <dependency>
             <groupId>org.apache.camel</groupId>
