+++ extensions/core-xml/deployment/src/main/java/org/apache/camel/quarkus/component/xml/deployment/XmlProcessor.java
@@ -16,14 +16,11 @@
  */
 package org.apache.camel.quarkus.component.xml.deployment;
 
-import io.quarkus.deployment.annotations.BuildProducer;
 import io.quarkus.deployment.annotations.BuildStep;
 import io.quarkus.deployment.annotations.ExecutionTime;
 import io.quarkus.deployment.annotations.Record;
 import io.quarkus.deployment.builditem.FeatureBuildItem;
-import io.quarkus.deployment.builditem.substrate.ReflectiveClassBuildItem;
 import io.quarkus.jaxb.deployment.JaxbFileRootBuildItem;
-import org.apache.camel.converter.jaxp.XmlConverter;
 import org.apache.camel.quarkus.component.xml.XmlRecorder;
 import org.apache.camel.quarkus.core.deployment.CamelModelJAXBContextFactoryBuildItem;
 import org.apache.camel.quarkus.core.deployment.CamelRoutesCollectorBuildItem;
@@ -67,23 +64,4 @@ class XmlProcessor {
     void initXmlReifiers(XmlRecorder recorder) {
         recorder.initXmlReifiers();
     }
-
-    @BuildStep
-    void reflective(BuildProducer<ReflectiveClassBuildItem> reflectiveClass) {
-        reflectiveClass.produce(new ReflectiveClassBuildItem(false, false,
-            "com.sun.org.apache.xerces.internal.jaxp.DocumentBuilderFactoryImpl",
-            "com.sun.org.apache.xalan.internal.xsltc.trax.TransformerFactoryImpl",
-            "com.sun.org.apache.xerces.internal.jaxp.SAXParserFactoryImpl",
-            "com.sun.xml.internal.stream.XMLInputFactoryImpl",
-            "com.sun.org.apache.xerces.internal.parsers.SAXParser",
-            XmlConverter.class.getCanonicalName()));
-
-        // javax.xml.namespace.QName is needed as it is used as part of the processor
-        // definitions in the DSL and parsers like Jackson (used in camel-k YAML DSL)
-        // fails if this class is cannot be instantiated reflectively.
-        reflectiveClass.produce(
-            new ReflectiveClassBuildItem(true, false, "javax.xml.namespace.QName")
-        );
-    }
-
 }
