+++ extensions/support/xml/deployment/pom.xml
@@ -20,13 +20,13 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
         <groupId>org.apache.camel.quarkus</groupId>
-        <artifactId>camel-quarkus-aws-s3-parent</artifactId>
+        <artifactId>camel-quarkus-support-xml-parent</artifactId>
         <version>0.2.1-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>camel-quarkus-aws-s3-deployment</artifactId>
-    <name>Camel Quarkus :: AWS S3 :: Deployment</name>
+    <artifactId>camel-quarkus-support-xml-deployment</artifactId>
+    <name>Camel Quarkus :: Support :: XML :: Deployment</name>
 
     <dependencyManagement>
         <dependencies>
@@ -42,16 +42,12 @@
 
     <dependencies>
         <dependency>
-            <groupId>org.apache.camel.quarkus</groupId>
-            <artifactId>camel-quarkus-core-deployment</artifactId>
+            <groupId>io.quarkus</groupId>
+            <artifactId>quarkus-core-deployment</artifactId>
         </dependency>
         <dependency>
             <groupId>org.apache.camel.quarkus</groupId>
-            <artifactId>camel-quarkus-support-xml-deployment</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.camel.quarkus</groupId>
-            <artifactId>camel-quarkus-aws-s3</artifactId>
+            <artifactId>camel-quarkus-support-xml</artifactId>
         </dependency>
     </dependencies>
