+++ extensions/support/xml/pom.xml
@@ -21,20 +21,19 @@
 
     <parent>
         <groupId>org.apache.camel.quarkus</groupId>
-        <artifactId>camel-quarkus-extensions</artifactId>
+        <artifactId>camel-quarkus-build-parent</artifactId>
         <version>0.2.1-SNAPSHOT</version>
+        <relativePath>../../../poms/build-parent/pom.xml</relativePath>
     </parent>
 
     <modelVersion>4.0.0</modelVersion>
 
-    <artifactId>camel-quarkus-extensions-support</artifactId>
-    <name>Camel Quarkus :: Support</name>
+    <artifactId>camel-quarkus-support-xml-parent</artifactId>
+    <name>Camel Quarkus :: Support :: XML</name>
     <packaging>pom</packaging>
-
     <modules>
-        <module>common</module>
-        <module>jetty</module>
-        <module>xstream</module>
-        <module>xml</module>
+        <module>deployment</module>
+        <module>runtime</module>
     </modules>
+
 </project>
