+++ poms/bom/pom.xml
@@ -268,6 +268,11 @@
                 <artifactId>camel-quarkus-support-xstream</artifactId>
                 <version>${camel-quarkus.version}</version>
             </dependency>
+            <dependency>
+                <groupId>org.apache.camel.quarkus</groupId>
+                <artifactId>camel-quarkus-support-xml</artifactId>
+                <version>${camel-quarkus.version}</version>
+            </dependency>
 
             <!-- Camel Quarkus Common (in alphabetical order) -->
             <dependency>
